selftests/bpf: Improve portability of unprivileged tests
The addition of general support for unprivileged tests in test_loader.c
breaks building test_verifier on non-glibc (e.g. musl) systems, due to the
inclusion of the glibc extension '<error.h>' in 'unpriv_helpers.c'.

Some review suggests usage of '<error.h>' under tools/ could be replaced
with '<err.h>', a BSD extension more widely supported (glibc and musl),
for better portability. However, in this case the header is actually not
needed, so remove it to restore building.

Fixes: 1d56ade032a4 ("selftests/bpf: Unprivileged tests for test_loader.c")
Signed-off-by: Tony Ambardar <Tony.Ambardar@gmail.com>