selftests/bpf: Improve bpf_iter/task_btf test debug
Previous commit [0] changed from using CHECK() to ASSERT() but left the
original long messages, which can be awkward to follow:

  do_btf_read:FAIL:check for btf representation of task_struct in iter data unexpected check for btf representation of task_struct in iter data: '(struct task_struct)' is not a substring of 'Raw BTF task'
  test_task_btf:FAIL:no task iteration, did BPF program run? unexpected no task iteration, did BPF program run?: actual 0 == expected 0

Also, if the ASSERT() on number of task iterations fails, it then skips an
ASSERT() that would check and print the relevant BPF prog return value.

Update the assertion messages to be consistent with other usage, and always
print the 'seq_err' return value on error.

0: f78625fdc95e ("selftests/bpf: Use ASSERT_* instead of CHECK")

Signed-off-by: Tony Ambardar <tony.ambardar@gmail.com>